Linux: Rename linux.envars to linux.envvars
